requirements.txt
psutil>=5.9.0
flask>=2.0.0
requests>=2.25.0
python-dotenv>=0.19.0